Fix stale connections, properties and new host
Fixed an error that caused connections to hang
resulting in an infinite wait for new
connections from the pool. Now the entity is
fully consumed, closed and discarded and the
returned data is the body of the response.

Updated the property configuration with test.

The API v2 has moved from beta.egree to
test.egree. Default hostname updated to this.