Update README, fix UTC and parameters handling
Updates the README to be more accurate regarding the receipt and usage
of culture.

Update all the Party and Case models to process non-UTC date times if
the client passes date times that are local. This is only for date times
that is client settable.

Added a class that process RequestParameters instead of hiding the
behavior inside the EgreeImpl.